Backend/pyproject.toml
[project]
name = "maat-backend"
version = "0.1.0"
description = "Deterministic core for rule-governed chess agent workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-chess>=1.999",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
